fix: align network_id type to u32 across AddNodeOpts, NodeConfig, and CLI (H3)
AddNodeOpts used u8 while NodeConfig used Option<u32>, requiring a
lossy cast. Align both to u32 for consistency and forward-compatibility.

+++ ant-cli/src/commands/node/add.rs
@@ -36,7 +36,7 @@ pub struct AddArgs {
 
     /// Network ID (default: 1 for mainnet)
     #[arg(long, default_value = "1")]
-    pub network_id: u8,
+    pub network_id: u32,
 
     /// Path to a local node binary
     #[arg(long, conflicts_with_all = &["version", "url"])]
